+++ src/main/java/com/common/POJO/ListNode.java
@@ -1,5 +1,9 @@
-package com.POJO;
+package com.common.POJO;
 
+/*
+* ListNode<E> : Defines a ListNode
+* E : The value type [String, Integer, Boolean]
+* */
 public class ListNode<E> {
     public E val;
     public ListNode<E> next;
