+++ src/main/java/com/dfs/DFSTree.java
@@ -1,8 +1,7 @@
 package com.dfs;
 
-import com.POJO.BinaryTree;
-import com.common.BinaryTreeUtils;
-import com.twoPointers.TPNodes;
+import com.common.POJO.BinaryTree;
+import com.common.Utils.BinaryTreeUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
